refactor: decompose code from functions.py to edit_functions.py; import in handlers.py; delete unused rows in remind_functions.py.
