+++ func/functions.py
@@ -131,122 +131,6 @@ def get_transactions_by_date(date, chat_id):
     return transactions
 
 
-def delete_transactions(chat_id, payment_id=None, recurrence_id=None):
-    conn = get_db_connection()
-    cursor = conn.cursor()
-    try:
-        if recurrence_id:
-            cursor.execute(
-                f"""
-                        UPDATE '{chat_id}' 
-                        SET is_active = 0 
-                        WHERE recurrence_id = ?
-                        """, (recurrence_id,)
-            )
-            conn.commit()
-
-        elif payment_id:
-            cursor.execute(
-                f"""
-                        UPDATE '{chat_id}' 
-                        SET is_active = 0 
-                        WHERE UUID = ?
-                        """, (payment_id,)
-            )
-            conn.commit()
-    except Exception as e:
-        logging.error(f"{e}")
-    finally:
-        conn.close()
-
-
-def undo_delete_transactions(chat_id, payment_id=None, recurrence_id=None, ):
-    conn = get_db_connection()
-    cursor = conn.cursor()
-    try:
-        if recurrence_id:
-            cursor.execute(
-                f"""
-                        UPDATE '{chat_id}' 
-                        SET is_active = 1 
-                        WHERE recurrence_id = ?
-                        """, (recurrence_id,)
-            )
-            conn.commit()
-
-        elif payment_id:
-            cursor.execute(
-                f"""
-                        UPDATE '{chat_id}' 
-                        SET is_active = 1 
-                        WHERE UUID = ?
-                        """, (payment_id,)
-            )
-            conn.commit()
-    except Exception as e:
-        logging.error(f"{e}")
-    finally:
-        conn.close()
-
-
-def done_transactions(chat_id, payment_uuid=None, recurrence_id=None):
-    conn = get_db_connection()
-    cursor = conn.cursor()
-    try:
-        if recurrence_id:
-            cursor.execute(
-                f"""
-                            UPDATE '{chat_id}' 
-                            SET execution_status = 1, is_active = 0
-                            WHERE recurrence_id = ?
-                            """, (recurrence_id,)
-            )
-            conn.commit()
-
-        elif payment_uuid:
-            cursor.execute(
-                f"""
-                            UPDATE '{chat_id}' 
-                            SET execution_status = 1, is_active = 0 
-                            WHERE UUID = ?
-                            """, (payment_uuid,)
-            )
-            conn.commit()
-    except Exception as e:
-        logging.error(f"{e}")
-    finally:
-        conn.close()
-
-
-def undone_transactions(chat_id, payment_uuid=None, recurrence_id=None):
-    conn = get_db_connection()
-    cursor = conn.cursor()
-    try:
-        if recurrence_id:
-            cursor.execute(
-                f"""
-                            UPDATE '{chat_id}' 
-                            SET execution_status = 0, is_active = 1
-                            WHERE recurrence_id = ?
-                            """, (recurrence_id,)
-            )
-            conn.commit()
-
-        elif payment_uuid:
-            cursor.execute(
-                f"""
-                            UPDATE '{chat_id}' 
-                            SET execution_status = 0, is_active = 1
-                            WHERE UUID = ?
-                            """, (payment_uuid,)
-            )
-            conn.commit()
-    except Exception as e:
-        logging.error(f"{e}")
-    finally:
-        conn.close()
-
-
 def get_transactions_in_date_range(start_date, end_date, message):
     chat_id = message.chat.id
     transaction_type_1 = db_transaction_types[1]
